requirements.txt
tqdm
rotary_embedding_torch
transformers
tokenizers
inflect
progressbar
einops
unidecode
scipy
librosa==0.9.1
torchaudio
openai-whisper
streamlit_mic_recorder
matplotlib==3.7
tensorflow-cpu==2.8.0
numpy==1.20
plotly
pandas
requests
opencv-python-headless
streamlit
uvicorn
torchaudio
plotly
azure-functions
fastapi
python-multipart
protobuf==3.20
